requirements.txt
allure-behave==2.9.43
allure-pytest==2.9.43
allure-python-commons==2.9.43
allure2-behave==2.3.4
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.1.0
behave==1.2.6
behave-html-formatter==0.9.8
behave2cucumber==1.0.3
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.4
decorator==4.4.1
gherkin-official==4.1.3
html-testRunner==1.1.2
idna==2.8
importlib-metadata==4.6.4
iniconfig==1.1.1
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.9.5
lxml==4.3.3
MarkupSafe==1.0
more-itertools==7.0.0
nose==1.3.7
packaging==21.0
parse==1.19.0
parse-type==0.5.2
parso==0.5.2
Paver==1.3.4
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
psutil==5.6.1
py==1.10.0
Pygments==2.5.2
pyparsing==2.4.7
pytest==6.2.4
requests==2.21.0
selenium==3.11.0
six==1.16.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
unittest-xml-reporting==2.2.0
urllib3==1.24.1
virtualenv==16.6.0
wcwidth==0.1.7
zipp==3.5.0
